+++ Pinecones/Hydra/src/main/java/com/pinecone/hydra/kernel/entity/KernelObject.java
@@ -3,8 +3,8 @@ package com.pinecone.hydra.kernel.entity;
 import com.pinecone.framework.util.id.GUID;
 
 public interface KernelObject {
-    int getEnumId();
-    void setEnumId(int id);
+    long getEnumId();
+    void setEnumId(long id);
     GUID getGuid();
     void setGuid(GUID guid);
     GUID getTargetGuid();
