+++ Pinecones/Hydra/src/main/java/com/pinecone/hydra/registry/entity/GenericNamespaceNodeMeta.java
@@ -4,25 +4,25 @@ import com.pinecone.framework.util.id.GUID;
 import com.pinecone.framework.util.json.hometype.BeanJSONEncoder;
 
 public class GenericNamespaceNodeMeta implements NamespaceNodeMeta{
-    private int enumId;
+    private long enumId;
 
     private GUID guid;
 
     public GenericNamespaceNodeMeta() {
     }
 
-    public GenericNamespaceNodeMeta( int enumId, GUID guid ) {
+    public GenericNamespaceNodeMeta( long enumId, GUID guid ) {
         this.enumId = enumId;
         this.guid = guid;
     }
 
     @Override
-    public int getEnumId() {
+    public long getEnumId() {
         return this.enumId;
     }
 
     @Override
-    public void setEnumId(int enumId) {
+    public void setEnumId(long enumId) {
         this.enumId = enumId;
     }
